api/requirements.txt
fastapi
uvicorn
sqlmodel
httpx
python-dotenv
pydantic-settings
libsql-experimental
sqlalchemy-libsql
